+++ HairCutAppAPI/DTOs/ComboDTOs/AdvancedGetCombosDTO.cs
@@ -23,8 +23,8 @@ namespace HairCutAppAPI.DTOs.ComboDTOs
         public string MaxCreatedDate { get; set; }
         public string MinLastUpdate { get; set; }
         public string MaxLastUpdate { get; set; }
-        public decimal MinPrice { get; set; }
-        public decimal MaxPrice { get; set; }
+        public long MinPrice { get; set; }
+        public long MaxPrice { get; set; }
         public int MinDuration { get; set; }
         public int MaxDuration { get; set; }
         public string SortBy { get; set; }
